requirements.txt
Appium-Python-Client>=4.0.0
pytest>=7.0.0
selenium>=4.0.0